Improve moving of mail templates
Use correct templating environment in RenderMailTemplatesCommand.
Remove unneccessary filter from mail templating environment (mail
templates don't link to asset files that need to be prefixed).
Replace anonymous UrlGenerator class with standalone one since it's used
in 3 places now.
Add new config settings to vagrant environment.
