+++ web/index.dev.php
@@ -18,7 +18,7 @@ use Monolog\Handler\BufferHandler;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use Symfony\Bridge\Twig\Extension\RoutingExtension;
-use WMDE\Fundraising\Frontend\Infrastructure\UrlGenerator;
+use WMDE\Fundraising\Frontend\App\UrlGenerator;
 
 /**
  * @var \WMDE\Fundraising\Frontend\Factories\FunFunFactory $ffFactory
@@ -118,18 +118,6 @@ $app['twig.loader.filesystem'] = $app->extend(
 
 $ffFactory->setSkinTwigEnvironment( $app['twig'] );
 
-$ffFactory->setUrlGenerator(
-	new class( $app['twig']->getExtension( RoutingExtension::class ) ) implements UrlGenerator {
-		private $routingExtension;
-
-		public function __construct( RoutingExtension $routingExtension ) {
-			$this->routingExtension = $routingExtension;
-		}
-
-		public function generateUrl( string $name, array $parameters = [] ): string {
-			return $this->routingExtension->getUrl( $name, $parameters );
-		}
-	}
-);
+$ffFactory->setUrlGenerator( new UrlGenerator( $app['twig']->getExtension( RoutingExtension::class ) ) );
 
 $app->run();
