+++ web/index.php
@@ -10,9 +10,9 @@ use Monolog\Handler\FingersCrossedHandler;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use Symfony\Bridge\Twig\Extension\RoutingExtension;
+use WMDE\Fundraising\Frontend\App\UrlGenerator;
 use WMDE\Fundraising\Frontend\Factories\FunFunFactory;
 use WMDE\Fundraising\Frontend\Infrastructure\ConfigReader;
-use WMDE\Fundraising\Frontend\Infrastructure\UrlGenerator;
 
 /**
  * @var FunFunFactory $ffFactory
@@ -84,18 +84,6 @@ $app = require __DIR__ . '/../app/bootstrap.php';
 
 $ffFactory->setSkinTwigEnvironment( $app['twig'] );
 
-$ffFactory->setUrlGenerator(
-	new class( $app['twig']->getExtension( RoutingExtension::class ) ) implements UrlGenerator {
-		private $routingExtension;
-
-		public function __construct( RoutingExtension $routingExtension ) {
-			$this->routingExtension = $routingExtension;
-		}
-
-		public function generateUrl( string $name, array $parameters = [] ): string {
-			return $this->routingExtension->getUrl( $name, $parameters );
-		}
-	}
-);
+$ffFactory->setUrlGenerator( new UrlGenerator( $app['twig']->getExtension( RoutingExtension::class ) ) );
 
 $app->run();
